+++ example/rails-51/config/initializers/bugsnag.rb
@@ -1,4 +1,4 @@
 Bugsnag.configure do |config|
   config.api_key = "YOUR_API_KEY"
-  config.track_sessions = true
+  config.auto_capture_sessions = true
 end