+++ lib/bugsnag.rb
@@ -39,8 +39,6 @@ module Bugsnag
         configuration.warn("No valid API key has been set, notifications will not be sent")
         @key_warning = true
       end
-
-      session_tracker.config = configuration
     end
 
     # Explicitly notify of an exception
@@ -116,8 +114,9 @@ module Bugsnag
 
         # Deliver
         configuration.info("Notifying #{configuration.endpoint} of #{report.exceptions.last[:errorClass]}")
-        options = {:headers => report.headers, :trim_payload => true}
-        Bugsnag::Delivery[configuration.delivery_method].deliver(configuration.endpoint, report.as_json, configuration, options)
+        options = {:headers => report.headers}
+        payload = ::JSON.dump(Bugsnag::Helpers.trim_if_needed(report.as_json))
+        Bugsnag::Delivery[configuration.delivery_method].deliver(configuration.endpoint, payload, configuration, options)
       end
     end
 
@@ -127,9 +126,14 @@ module Bugsnag
       @configuration || LOCK.synchronize { @configuration ||= Bugsnag::Configuration.new }
     end
 
+    # Session tracking
     def session_tracker
       @session_tracker = nil unless defined?(@session_tracker)
-      @session_tracker || LOCK.synchronize { @session_tracker ||= Bugsnag::SessionTracker.new(configuration)}
+      @session_tracker || LOCK.synchronize { @session_tracker ||= Bugsnag::SessionTracker.new}
+    end
+
+    def start_session
+      session_tracker.start_session
     end
 
     # Allow access to "before notify" callbacks